ISSUE-2: VMEmulator built-in classes not useable from command line
The use of built-in classes is now controlled by the environment
variable N2T_VM_USE_BUILTINS.

If set to "yes" (case insesitive), the built-in classes will be
used without prompting.

If set to "no", the built-in classes won't be used (again, no
prompting).

In any other case, we'll fallback to the previous behaviour. That
is, prompt for GUI and reject for CLI.